requirements.txt
cffi==1.13.2
cryptography==2.8
pycparser==2.19
pycrypto==2.6.1
six==1.13.0
